aether/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.12.1

# Database drivers
psycopg2-binary==2.9.9
aiosqlite==0.19.0

# Vector store
chromadb==0.4.18

# AI and ML
openai==1.3.7
langchain==0.0.340
sentence-transformers==2.2.2
numpy==1.24.4

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pyjwt==2.8.0
bcrypt==4.1.2
slowapi==0.1.9

# Configuration
python-dotenv==1.0.0

# Logging and monitoring
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Google integrations
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3